+++ src/sep.ira/calculator.maui/docs/changelog.md
@@ -2,7 +2,10 @@
 
 All notable changes to this Maui application will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-[1.0.9666]: https://www.github.com/atecoder/dn.data.finance
+[1.0.9673]: https://www.github.com/atecoder/dn.data.finance
+
+## [1.0.9673] - 2026-06-26
+- Updated.
 
 ## [1.0.9666] - 2026-06-19
 - Created