+++ src/sep.ira/calculator.xunits/docs/changelog.md
@@ -2,10 +2,11 @@
 
 All notable changes to this unit tests application will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-[1.0.9666]: https://www.github.com/atecoder/dn.data.finance
+[1.0.9673]: https://www.github.com/atecoder/dn.data.finance
 
-## [1.0.9666] - 2026-06-19
+## [1.0.9673] - 2026-06-26
 - Define financial values as Decimal.
+- Run all tests and document.
 
 ## [1.0.9664] - 2026-06-17
 - Add unit tests for the appreciator calculations.